Mangle Include
This add support to import prefabs and automatically set them to the
slope on which the player is standing on you can also do this single
objects /mb command. I have not touched this stuff in ages so some stuff
needs to be deleted and it needs to be test but pretty sure to should
work.