Use databases for 'allobjects' and 'modelsizes'.
Both are now faster in game (I didn't expect them to be but now '/gtaobjects' take merely two seconds at most). Both also reduce compile time, but it still has to compress a lot so it's not reduced as much as I hoped (not a bad thing, it reduced it drastically).

This resolves #19.