Make a working /gaexport.
Since the file isn't being created for some reason (I've tried multiple ways of creating the file: ftouch from YSI, all of the file modes except io_reading since it's the only one that doesn't automatically create the file, and creating it beforehand which is obviously useless), I've resorted to a temporary fix. Using the blank.txt file that is already included so that GitHub saves the folder.

Again, this file method is only temporary.